Moves some misc traits/activities to Mods.Common, moves some RA traits into the Traits folder
