+++ OpenRA.Mods.RA/Scripting/Properties/TransformProperties.cs
@@ -8,6 +8,7 @@
  */
 #endregion
 
+using OpenRA.Mods.Common.Traits;
 using OpenRA.Mods.RA.Traits;
 using OpenRA.Scripting;
 using OpenRA.Traits;
