+++ OpenRA.Mods.RA/Traits/Disguise.cs
