+++ OpenRA.Mods.RA/Traits/EjectOnDeath.cs
