+++ OpenRA.Mods.RA/Traits/ParaDrop.cs
@@ -13,10 +13,9 @@ using System.Collections.Generic;
 using OpenRA.Mods.Common.Traits;
 using OpenRA.Mods.RA.Activities;
 using OpenRA.Mods.RA.Effects;
-using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	[Desc("This unit can spawn and eject other actors while flying.")]
 	public class ParaDropInfo : ITraitInfo, Requires<CargoInfo>
