+++ OpenRA.Mods.RA/Traits/TransformOnPassenger.cs
@@ -13,7 +13,7 @@ using OpenRA.Mods.Common.Activities;
 using OpenRA.Mods.Common.Traits;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	public class TransformOnPassengerInfo : ITraitInfo
 	{
